requirements.txt
arrow==1.2.3
asgiref==3.7.2
autopep8==2.0.4
boto3==1.28.52
botocore==1.31.52
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
dateparser==1.1.8
Django==4.2.4
django-cors-headers==4.9.0
django-eb-sqs==1.44
django-ses==3.5.0
django-silk==5.0.3
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
googlemaps==4.10.0
gprof2dot==2022.7.29
gunicorn==23.0.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
moment==0.12.1
mysqlclient==2.2.0
openrouteservice==2.3.3
packaging==25.0
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
setuptools
regex==2023.8.8
requests==2.31.0
s3transfer==0.6.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
times==0.7
tomli==2.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16

